♻️ / #145 - AlertPage template로 분리
